Simplified the logic of insertions and deletions
Details
-------
Previously, each insertion/deletion trial had a 'mode_forward' to keep track of what was being inserted or deleted, and a ‘move_backward’ so that the move could be reversed if the trial was rejected. Now, there is only one ‘mode’, with the reversal of trial upon rejection greatly simplified. This should make the code easier to read.