Fix bug pertaining to indexing in Swapper.update_fractional_ion
Why:

A bug existed in Swapper.update_fractional_atom that caused a TypeError
during runtime.

The final_force array wasn't indexed by atom, but by parameter name.
This lead to a TypeError at runtime because instead it was indexed with
a the string "charge", instead of an index such as atm_index.

What:
This commit adds indexing final_force by atm_index to the relevant lines
 in Swapper.